Moved AkkaSource type to sdk package object
