+++ delta/sdk/src/main/scala/ch/epfl/bluebrain/nexus/delta/sdk/package.scala
@@ -1,5 +1,7 @@
 package ch.epfl.bluebrain.nexus.delta
 
+import akka.stream.scaladsl.Source
+import akka.util.ByteString
 import ch.epfl.bluebrain.nexus.delta.sdk.model.{ResourceF, ResourceRef}
 import ch.epfl.bluebrain.nexus.delta.sdk.model.acls.Acl
 import ch.epfl.bluebrain.nexus.delta.sdk.model.identities.Caller
@@ -58,5 +60,7 @@ package object sdk {
     * Type alias for resolver resolution
     */
   type Resolve[A] = (ResourceRef, ProjectRef, Caller) => IO[ResourceResolutionReport, A]
+
+  type AkkaSource = Source[ByteString, Any]
 
 }
